requirements.txt
colorama==0.4.6
numpy==2.1.3
opencv-python==4.9.0.80
pillow==11.0.0
PyYAML==6.0.2
tqdm==4.67.0
